# src/main/resources/api.yml
openapi: 3.1.0
info:
  version: v1
  title: content-vault API
  description: content-vault RESTFul api for openapi 3.1
  termsOfService: 'https://jumkid.com/terms-of-use'
  contact:
    name: Chooli Yip
    url: jumkid.com
    email: chooli.yip@gmail.com
  license:
    name: Jumkid Public License
    url: 'http://license.jumkid.com'
servers:
  - description: Testing
    url: 'http://test.jumkid.com'
  - description: Live production
    url: 'http://api.jumkid.com'
...
security:
  - bearerAuth: []
paths:
...
  '/content/{id}':
    get:
      tags:
        - v1
      operationId: getPlain
      description: get plain content
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
        - name: ignoreTitle
          in: query
          required: false
          description: if the response body should ignore the title
          schema:
            type: boolean
      responses:
        '200':
          description: successfully fetch content in plain format
          content:
            test/plain:
              schema:
                type: string
  /content:
    post:
      tags:
        - v1
      operationId: addTextContent
      description: save a new plain content
      parameters:
        - name: title
          in: query
          schema:
            type: string
        - name: content
          in: query
          schema:
            type: string
        - name: accessScope
          in: query
          required: true
          schema:
            $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
      summary: Add text content
  /content/html:
    post:
      tags:
        - v1
      operationId: addHtmlContent
      description: save a new text content
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
        - name: accessScope
          in: query
          required: true
          schema:
            $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope    
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
      summary: Add html content
  '/content/stream/{id}':
    get:
      tags:
        - v1
      operationId: stream
      description: stream content
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
        - name: size
          in: query
          description: size option
          schema:
            $ref: '#/components/schemas/ThumbnailSize'
      responses:
        '200':
          description: binary stream as response
  '/content/thumbnail/{id}':
    get:
      tags:
        - v1
      operationId: thumbnail
      description: get thumbnail image by given content id
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      responses:
        '200':
          description: binary as response

## Gallery API
...
components:
  schemas:
    MediaFile:
      description: Model of media file
      properties:
        id:
          type: string
          example: 1mWG-nQBWmmN0po1BUoG
        filename:
          type: string
          example: foo.txt
        mimeType:
          type: string
          example: text/plain
        size:
          type: integer
          description: file size in byte
        title:
          type: string
        content:
          type: string
        activated:
          type: boolean
        accessScope:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope
        createdBy:
          type: string
        creationDate:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
        modifiedBy:
          type: string
        modifiedOn:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
        props:
          type: array
          items:
            description: list of additional properties
            properties:
              name:
                type: string
              textValue:
                type: string
              dateValue:
                $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
              numberValue:
                type: integer
        tags:
          type: array
          items:
            description: list of tags
            type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/MediaFile'
...
    ThumbnailSize:
      type: string
      enum:
        - small
        - large
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT